services/worker/requirements.txt
psycopg2-binary==2.9.9
redis==5.0.4
prometheus-client==0.20.0
